requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.29
pydantic==2.6.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.32.3
